refactor: move dimensions to theme
