+++ src/components/Tabs/TabsNav.js
@@ -2,8 +2,8 @@
 import { jsx } from "@emotion/core"
 import PropTypes from "prop-types"
 
-import { dimensions } from "../../utils/presets"
 import breakpoints from "../../theme/breakpoints"
+import dimensions from "../../theme/dimensions"
 
 const TabsNav = ({ children }) => (
   <nav
