+++ src/components/Toast/Toast.tsx
@@ -3,9 +3,10 @@ import React from "react"
 import Alert from "@reach/alert"
 import { keyframes, css, jsx } from "@emotion/core"
 import { MdDone, MdClose, MdWarning } from "react-icons/md"
-import { dimensions, radius } from "../../utils/presets"
+import { radius } from "../../utils/presets"
 import fontSizes from "../../theme/fontSizes"
 import { ToastTone } from "./types"
+import dimensions from "../../theme/dimensions"
 import space from "../../theme/space"
 import colors from "../../theme/colors"
