+++ src/theme/dimensions.ts
@@ -1,0 +1,28 @@
+const dimensions = {
+  siteHeader: {
+    height: `3.75rem`,
+  },
+  layout: {
+    width: 90,
+  },
+  dropdown: {
+    list: {
+      width: `250px`,
+    },
+  },
+  toast: {
+    minHeight: `3rem`,
+  },
+  pageNav: {
+    minHeight: `5rem`,
+  },
+  pagePadding: {
+    mobile: `1.4rem`,
+    tablet: `2rem`,
+  },
+  buildList: {
+    indentation: `5rem`,
+  },
+}
+
+export default dimensions
