+++ src/utils/presets/index.ts
@@ -1,4 +1,4 @@
 export { fontFamilies } from "./typography"
-export { dimensions, radius } from "./sizes"
+export { radius } from "./sizes"
 export { zIndices } from "./zIndices"
 export { transition } from "./transition"
