+++ src/utils/presets/sizes.ts
@@ -2,30 +2,3 @@ export const radius = {
   default: `4px`,
   large: `8px`,
 }
-
-export const dimensions = {
-  siteHeader: {
-    height: `3.75rem`,
-  },
-  layout: {
-    width: 90,
-  },
-  dropdown: {
-    list: {
-      width: `250px`,
-    },
-  },
-  toast: {
-    minHeight: `3rem`,
-  },
-  pageNav: {
-    minHeight: `5rem`,
-  },
-  pagePadding: {
-    mobile: `1.4rem`,
-    tablet: `2rem`,
-  },
-  buildList: {
-    indentation: `5rem`,
-  },
-}
